+++ src/fa/maps.py
@@ -495,52 +495,6 @@ def preview(mapname, pixmap=False):
         logger.error("Map Preview Exception", exc_info=sys.exc_info())
 
 
-
-class Downloader(QtCore.QObject):
-    progress_reset = QtCore.pyqtSignal()
-    progress_value = QtCore.pyqtSignal(int)
-    progress_maximum = QtCore.pyqtSignal(int)
-    progress_log = QtCore.pyqtSignal(str)
-    failed = QtCore.pyqtSignal(str)
-    finished = QtCore.pyqtSignal()
-
-    def __init__(self, map_name, download_root=VAULT_DOWNLOAD_ROOT, parent=None):
-        QtCore.QObject.__init__(self, parent)
-        self.map_name = map_name
-        self.url = QtCore.QUrl(download_root + name2link(map_name))
-        self.data = QtCore.QByteArray()
-        self.reply = None
-
-    @QtCore.pyqtSlot(int, int)
-    def _propagate_progress(self, value, maximum):
-        self.progress_value.emit(value)
-        self.progress_maximum.emit(maximum)
-
-    @QtCore.pyqtSlot()
-    def _read_data(self):
-        self.data.append(self.reply.readAll())
-
-    @QtCore.pyqtSlot()
-    def _save_data(self):
-
-        self.finished.emit()
-
-    @QtCore.pyqtSlot()
-    def abort(self):
-        self.reply.close()
-        self.failed.emit()
-
-    @QtCore.pyqtSlot()
-    def run(self):
-        self.progress_reset.emit()
-        self.progress_log.emit("Downloading map " + self.map_name)
-        self.reply = util.network.get(QtNetwork.QNetworkRequest(self.url))
-        self.reply.readyRead.connect(self._read_data)
-        self.reply.finished.connect(self._save_data)
-        self.reply.downloadProgress.connect(self._propagate_progress)
-
-
-
 def downloadMap(name, silent=False):
     '''
     Download a map from the vault with the given name